Enabling non-hardcoded run of our test infrastructure (#145)
As per issue https://github.com/tenstorrent/tt-xla/issues/9, our test
infra had a problem when detecting two TT devices, as is the case on the
two-chip N300, which was workarounded by hardcoding which devices to
use. This PR fixes that, instead just setting that the number of devices
that the backend uses is capped to 1. Additionally, there was a bug in
our old infra which ran the tests on cpu instead of a TT device, so this
PR also fixes that.

Fixes https://github.com/tenstorrent/tt-xla/issues/9.