Adding infrastructure for multichip tests (#231)
### Ticket
Part of issue https://github.com/tenstorrent/tt-xla/issues/211

### Problem description
Integrating new tester class, called `MultichipTester`, into our testing
infrastructure in`tt-xla`. It will receive the input and output
shardings from the user, the function to be tested, and the golden
function which will run on CPU, in a non-sharded mode. It runs the
received data and compared the results.