Adding implementation for PJRT_Executable_OutputElementTypes and PJRT_Executable_OutputDimensions (#238)
### Ticket
https://github.com/tenstorrent/tt-xla/issues/164

### Problem description
`PJRT_Executable_OutputElementTypes` and
`PJRT_Executable_OutputDimensions` were unimplemented.

### What's changed
In `LoadedExecutableInstance::Execute` we got the shape for every
`PJRT_Buffer` output from `binary.getProgramOutputs` and
`image_->isOutputScalar`. Now we store the shape in `ExecutableImage`
class and have a getter for the final shape, and not `isOutputScalar`.

Adding a vector that stores `PJRT_Buffer_Type` for every output in
`ExecutableImage`, as well as a getter, for the implementation of
`PJRT_Executable_OutputElementTypes`.

`PJRT_Executable_OutputDimensions` requires the dimensions to be in a
different format. It returns an array of output dimension sizes and an
array of concatenated dimensions, so we also store that type of pointer
in `ExecutableImage`. We store it because we are responsible for memory
deallocation, and not the caller of `PJRT_Executable_OutputDimensions`.
But because concatenated dimensions are rarely needed, these fields are
`nullptr` until `get_output_dims_concatenated` is called.

### Notes
It is possible that `PJRT_Executable_OutputElementTypes` should return
only different unique PJRT_Buffer_Types, and not a type for every
output.

We might want to store information about stride of every output in
`ExecutableImage` as well, so `binary.getProgramOutput` wouldn't have to
be called in `LoadedExecutableInstance::Execute`, and all the
information about outputs would be at one place.

### Checklist
- [ ] New/Existing tests provide coverage for changes