Changing the multichip testing infrastructure to support virtualized CPU meshes (#352)
Currently in the multichip testing infra, we are required the writer of
the tests to write a separate, single device golden function for
testing. This PR adds support for virtualizing cpu into multiple devices
and making a CPU mesh, so only the function that is tested needs to be
written.