Einnes tt/xla readme (#623)
### Ticket
#606 

### Problem description
As part of standardizing forge repos, the readme is redesigned to match
the type of info we want there. All details about building the project
are still available, but now on a getting started page. As we add
content, particularly if there's build vs. getting started pages, the
pages can be standardized further.

### What's changed
Describe the approach used to solve the problem.
More uniform appearance, with information being provided in a way that
will be consistent with the other repos. (The other repos aren't done
yet, so it will be a few cycles to get everything matching.)

### Checklist
- [ ] New/Existing tests provide coverage for changes

---------

Co-authored-by: Marko Rakita <mrakita@tenstorrent.com>