Fix CI permissions issue (#629)
### Ticket
/

### Problem description
Github actions have started throwing an error about missing permissions
for the build-and-test workflow.
Just as a heads up all of the "on PR", "on push" and "on nightly"
workflows will be failing until this is resolved.

### What's changed
Removing the specified permissions in the build-and-test workflow
resolves the issue.

### Checklist
- [ ] On PR workflow is passing

<img width="1542" alt="image"
src="https://github.com/user-attachments/assets/ddd9530d-72ca-4580-9832-78d5ada05d6e"
/>