Fix permissions in CI issue (explicitly set permissions) (#632)
### Ticket
/

### Problem description
The current approach of giving permissions to the `build-and-test`
workflow would not work in the case of open source contributions.
Github doesn't allow the secrets to be pulled into the workflow if it is
triggered from an external fork.

### What's changed
Changed the approach for solving the permissions issue in the CI.
We now explicitly give permissions to the `build-and-test` workflow
instead of using the custom token (`GH_TOKEN`).

### Checklist
- [ ] On PR workflow is passing