Release artifact generation and workflow speed up (#645)
### Ticket
/

### Problem description
Tt-xla didn't have a setup for releasing a wheel, since the setup was
created recently we want to generate it in the workflow.

### What's changed
`on pr`, `on push` and `on nightly` workflows now generate a
`xla-whl-release` artifact.

Additionaly: 
In the `on pr`, `on push` and `on nightly` workflows, tt-xla is now
built only once and then reused in all the test runs.
Workflows are now faster because of it.

### Checklist
- [ ] On PR workflow is passing

#### Future improvement
This PR addresses build artifact reuse and wheel generation.
There will be a PR for a cleaner refactor of the `on nightly` workflow
with some other improvements.

---------

Co-authored-by: Kristijan Mitrovic <kmitrovic@tenstorrent.com>