feat: Add tag "latest" to the newest docker image pushed from main (#677)
### Ticket
/

### Problem description
The "latest" tag doesn't exist for tt-xla's docker images.

### What's changed
Every docker image build process on main will add the tag `latest` to
the docker image in it, so the newest image built on main will have the
tag `latest`.
`Latest` tag can now be used to pull the tt-xla's newest docker image.
Tag is added for convenience and because a similar thing exists for
other frontends.

### Checklist
- [ ] New/Existing tests provide coverage for changes