Fixing the return of replicated scalars (#683)
In the new jax 0.6.0, if we are returning a replicated scalar value in a
multidevice environment, the StableHLO graph will annotate it as a
single-device value, therefore out runtime will produce a single device
tensor as output. To make it compatible with PJRT, we need to return
this value as output for each device in the multi device environment,
which this PR does.