Added support for torch in test infrastructure (#573)
### Problem description
We needed a way to support writing Torch models in our tests. Thanks to
[torch_xla](https://github.com/pytorch/xla) Torch can hook to our custom
PJRT plugin, as we did with JAX. What was left is to accommodate test
infra to support different paths for jax and torch.

Result of this effort is that writing torch tests for ops and graphs now
looks exactly the same as it was for jax, while writing a model test
requires user to inherit `TorchModelTester`, while jax tests have been
switched to inherit `JaxModelTester` (or `JaxMultichipModelTester` for
multichip models. There is no support for multichip for torch.).

Test writer needs to write a model in `torch` or `jax`, write custom
tester by inheriting `TorchModelTester` or `JaxModelTester`, override
some methods and that's it. Tester will take care of all the complexity
of compiling the model, running it on device and comparing the results
with CPU.

### What's changed
- Due to the need to support both jax and torch and different ways of
doing stuff in those frameworks, this is what class inheritance diagram
looks like now

<img width="2129" alt="image"
src="https://github.com/user-attachments/assets/0d3b10d9-d1db-44f9-8d85-e439edcdde70"
/>

- Here is the new folder structure:
```
tests/infra/
|-- __init__.py
|-- comparators
|   |-- __init__.py
|   |-- comparator.py
|   |-- comparator_factory.py
|   |-- comparison_config.py
|   |-- jax_comparator.py
|   `-- torch_comparator.py
|-- connectors
|   |-- __init__.py
|   |-- device_connector.py
|   |-- device_connector_factory.py
|   |-- jax_device_connector.py
|   `-- torch_device_connector.py
|-- runners
|   |-- __init__.py
|   |-- device_runner.py
|   |-- device_runner_factory.py
|   |-- jax_device_runner.py
|   |-- torch_device_runner.py
|   `-- utils.py
|-- testers
|   |-- __init__.py
|   |-- base_tester.py
|   |-- multichip
|   |   |-- __init__.py
|   |   |-- graph
|   |   |   |-- __init__.py
|   |   |   `-- jax_multichip_graph_tester.py
|   |   |-- model
|   |   |   |-- __init__.py
|   |   |   `-- jax_multichip_model_tester.py
|   |   `-- op
|   |       |-- __init__.py
|   |       `-- jax_multichip_op_tester.py
|   `-- single_chip
|       |-- __init__.py
|       |-- graph
|       |   |-- __init__.py
|       |   `-- graph_tester.py
|       |-- model
|       |   |-- __init__.py
|       |   |-- jax_model_tester.py
|       |   |-- model_tester.py
|       |   `-- torch_model_tester.py
|       `-- op
|           |-- __init__.py
|           `-- op_tester.py
|-- utilities
|   |-- __init__.py
|   |-- jax_multichip_utils.py
|   |-- types.py
|   `-- utils.py
`-- workloads
    |-- __init__.py
    |-- jax_workload.py
    |-- torch_workload.py
    |-- workload.py
    `-- workload_factory.py
```

- We now have a couple of "main" `*Tester` classes and different helper
classes which testers use kinda like
[strategies](https://refactoring.guru/design-patterns/strategy) on how
to do certain things. Strategy classes are meant to encapsulate a
certain part of the process (like connecting to a device, running a
workload on it, comparing the results, etc), and are thus kept small.
They all have a factory class, base class and `Jax` and `Torch`
subclasses.
- `comparators` encapsulate differences in comparing tensors in torch
and jax (simply put, in case of jax tensors, we have to run `jax`
commands, in case of torch tensors we have to run `torch` commands, so
this class captures that in order not to have a bunch of dual util
functions like `compare_pcc_torch` and `compare_pcc_jax`)
- `connectors` take care of connecting to TT device and CPU.
`DeviceConnector` class is devised as thread-safe singleton with simple
public `connect*` methods. All the complexity of registering PJRT plugin
and implementing the `connect*` methods is left to subclasses.
- Both connector classes have global instance variable exposed in their
respective modules. This instance is used to ensure connectors are the
first thing that is instantiated when running tests through
`conftest.py::initialize_device_connectors`. Inheritance chain and infra
API exposed to the test writer use `DeviceConnectorFactory` to fetch
those global instances of connectors (and initialize them if not
already). This makes sure plugins are registered and devices are
connected before any other `jax` or `torch` commands are executed.
- `runners` encapsulate running something on some device. That
"something" I call `Workload` (it itself is encapsulated in
`tests/infra/utilities/workloads`). Runners hide away the complexity of
sending workload to device and executing it there. Take note that no
support for multichip has been added to torch runner.
- `workloads` are also split into jax and torch variants
- `testers` are now split into `single_chip` and `multichip` dirs (to
logically match how tests are split, even though the inheritance
structure isn't that strict and is kept loose in case of future changes)
- `OpTester` exposes `run_op_test` and `run_op_test_with_random_inputs`
which is all that's needed to write an op test. Similar for
`GraphTester` which inherits it.
- Multichip testers exist only for JAX for now. `JaxMultichipOpTester`
exposes `run_jax_multichip_op_test_with_random_inputs`, while
`JaxMultichipGraphTester` exposes
`run_jax_multichip_graph_test_with_random_inputs`, similar to single
chip testers above.
- `JaxModelTester`, `JaxMultichipModelTester` and `TorchModelTester` are
meant to be inherited by custom testers and appropriate methods
overloaded. Calling public `test()` method on any tester will trigger
the entire logic of sending workload to device and executing it there,
and afterwards comparing the results with CPU.
- Added two examples in `examples/torch` on how to register our PJRT
plugin and write a simple torch model and run it on our device
- Added short public docs which explains how infra works
- Set `torch`, `torch_xla` and `torchvision` to most up to date versions
in `requirements.txt`
- All tests updated to use new infra API

### Checklist
- [x] New/Existing tests provide coverage for changes