Fix incorrect wheel file path in getting_started_build_from_source.md (#770)
### Problem description
The original path to the .whl file in
getting_started_build_from_source.md was incorrect. It was already
corrected in [this PR](https://github.com/tenstorrent/tt-xla/pull/720)
in the Getting Started section, but it is now incorrect again in the
getting_started_build_from_source.md file.

### What's changed
Updated the wheel file path in the documentation to the correct location