Split test execution in CI (#816)
### Ticket
I have added support for splitting the execution of tests in the CI.
On nightly workflow will split the tests on 3 runners by default, but
the dispatch of the on nightly can use a different number of runners
1,2,3,4 and 8.
The test suite that will currently be split is only the
`./tests/jax/multi_chip/n300`. Although this can easily be changed and
expanded.

### Problem description
Provide context for the problem.

### What's changed
Describe the approach used to solve the problem.
Summarize the changes made and its impact.

### Checklist
- [ ] New/Existing tests provide coverage for changes

Note:
Here is the passing nightly
[workflow](https://github.com/tenstorrent/tt-xla/actions/runs/16288578620)
with the newest change. Only main rebase was done after it.