Update Autoencoder postprocess to use PIL (#69)
### Ticket
As stated in this [PR](https://github.com/tenstorrent/tt-xla/pull/847)
removing the usage of Matplotlib and used PIL instead.

### Problem description
Postprocessing function of `Autoencoder` model uses `Matplotlib.pyplot`
to save the image which is a larger library in size. In order to avoid
using such a larger package, `PIL` is used to save the image