[build] wheel restructuring to handle both frameworks
Restructures the wheel to use a shared PJRT package with thin
framework-specific wrappers, eliminating binary duplication and
improving maintainability. The framework specific wrappers handle plugin
registration. Torch XLA plugin registration is implemented.

The wheel structure looks like this:
  - `pjrt_plugin_tt` package contains the PJRT device binary and
    `tt-mlir` installation
  - thin `jax_plugin_tt` and `torch_plugin_tt` wrappers import from shared
    package.

The `setup.py` is simplified - custom logic for copying the needed
binaries is removed and `setup.py` now performs the `cmake` build &
install so that the needed binaries are placed where `setuptools`
expects it.

Build process is also improved so that we use editable installations.
Now, when building the project via `cmake --build build`, the python
packages are installed in the `venv`. This way we no longer need any
logic for finding the plugins in the test infra - since it doesn't
matter if the user installed us as a wheel or as an editable
installation (framework specific wrappers should handle that).

For torch xla registration, we are able to automatically present our
selves to the framework. However, the user still needs to manually
choose our device; i.e. `torch_xla.runtime.set_device_type('TT')`.

Closes #986
Closes #988