add test for monodepth2 model (#818)
### Problem description
- Attempted to add a test for the Monodepth2 PyTorch model using the
tt-forge-models repository. However, the model was missing a loading
function and did not automatically download the required .pth file. This
caused import and test failures.
- To address this, I added the missing model download logic and opened a
[PR ](https://github.com/tenstorrent/tt-forge-models/pull/40)in the
tt-forge-models repository. I will keep this PR is currently in draft
and will remain so until the upstream changes are merged.

### What's changed
Added test files to test monodepth2 model