Improve docker build wait time (#1129)
### Ticket
N/A

### Problem description
We are using same machine to check if we need to build docker and
perform the build. This can lead to long queue times waiting for builder
machine, and in majority of runs we don't need to build the image.

### What's changed
Use ubuntu-latest which are always available to check if we need to
build docker.

### Checklist
- [x] New/Existing tests provide coverage for changes