Enable lower data formats (bfloat16 and bfp8) through Jax (#1102)
### Ticket


### Problem description


### What's changed
- Add CompilerConfig and propagate it to compile methods in tester
classes
- Add new compile option enable_bfp8_conversion that enables block fp8
in MLIR. For this to work TTIR graph has to be in bfloat16. Final graph
will have input and output nodes in bfloat16 and everything else in
bfp8. Essentially adding type casts at the beginning and in the end of
the graph, while all intermediate results are in bfp8. This bfloat16
wrapping is done because block formats are TT hardware specific, and
user should provide and get tensors of common dtype.
- Make all compile methods void - they change workload object inplace so
they don't have to return it. Change is in all tester classes: OpTester,
ModelTester, JaxModelTester, JaxMultichipModelTester, TorchModelTester,
JaxMultichipOpTester
- Add lower dataformat tests (bfloat16 and bfp8) for jax ops that have
float inputs.