Add latest tag to the tt-xla docker ird image as well (#1262)
### Ticket
/

### Problem description
The `latest` tag was only being added to the `ci` variant of the tt-xla
Docker image.

### What's changed
The job which adds the `latest` tag is now expanded to also add the tag
to the `ird` variant of the Docker image.

Note: I have also added myself as a CODEOWNER for the `/.github`
directory.

### Checklist
- [ ] New/Existing tests provide coverage for changes