Deleting the destructor in JaxModelTester (#1282)
Currently we have a destructor which deletes the huggingface files from
the cache when the model test has finished running. This cache is shared
with other CI runners and as such, can influence their tests.
Additionally, this does not relieve pressure on the disk, as huggingface
cache is not stored in local disk space of a runner. Becuse of these
reasons, deleting that destructor.