reuse device for both single & multi chip
We open the device when creating the `ClientInstance`. Then when
executing the models, in `LoadedExecutableInstance::execute()`, we
verify the wanted mesh shape for the device. If the device mesh shape
matches the target mesh shape, we simply reuse the previously opened
mesh device. In case we need a different mesh shape, we will close and
reopen the mesh device with wanted mesh shape.

There were issues with using sub-meshes and also with reshaping mesh
device. Hence, i've decided to go with this simples approach (for now).