[pjrt] reuse device if possible
We are currently opening/closing the device on each
`LoadedExecutableInstance::execute()` call.

This change enables us to open the device in advance and reuse it, for
example across multiple executions of the same model. This will allow us
to leave tensors (weights/constants) on the device, improving the
performance on subsequent executions - avoid constantly transferring
tensors from host to device.

We open the device when creating the `ClientInstance`. Then when
executing the models, in `LoadedExecutableInstance::execute()`, we
verify the wanted mesh shape for the device. If the device mesh shape
matches the target mesh shape, we simply reuse the previously opened
mesh device. In case we need a different mesh shape, we will close and
reopen the mesh device with wanted mesh shape.

There were issues with using sub-meshes and also with reshaping mesh
device. Hence, I've decided to go with the simplest approach which
works.

Issue #952
Closes #1415