Rebuild docker image on requirements change (#1349)
### Ticket
Closes #1324 

### Problem description
Using pre-installed python packages was causing problems with forge
model tests, as they required older version of transformers lib than it
was installed in docker, and pip is not downgrading libraries in this
case.
Use of system packages was disabled until we find solution for this.

### What's changed
We now trigger docker build on change to requirements files, which will
ensure same versions get installed.
Enable again use of system packages

### Checklist
- [x] New/Existing tests provide coverage for changes