update torch-xla install to use tt's fork wheel (#1320)
### Ticket
#1322 

### Problem description
We have our own fork for torch-xla that we haven't been able to use on
tt-xla due to conflicts with jax requirements. We have uplifted the
tenstorrent/torch-xla wheel to use python3.11 and jax==0.7.1 and we want
to update tt-xla's requirements to use this new wheel to align with all
tt front ends.

### What's changed
Updated requirements.txt to install tenstorrent/torch-xla latest wheel

### Checklist
- [x] New/Existing tests provide coverage for changes