Serialize IRs alongside flatbuffers (#947)
### Ticket
closes #932 

### Problem description
Perf team needs TTIR and TTNN MLIR code to be serialized alongside
flatbuffer binaries.

### What's changed
The compile logic was changed to store MLIR code as necessary and pass
it on, all the way to SerializedExecutableInstance class.
Serialization format was changed from just the fb to a custom format
that incorporates the needed IRs alongside the fb.
Python side helper function was refactored to work with the new
requirements.

### Checklist
- [ ] New/Existing tests provide coverage for changes