[ci] run examples in ci
Pytest wrapper test which runs python scripts in the `examples/`
directory is implemented. The test is parameterized by all collected python
scripts. Example scripts are ran as subprocesses and their exit code is
checked for failure.

Basic tests preset is extended to run this new test. All examples are
ran on `n300` so that both single and multi-chip scripts can work. This
can be expanded in the future if we need more sophisticated setups.

Side note: had to lower the number of iterations in
`simple_regression.py` in order for tests not to take too long.

Closes #137