add --pre flag info (#1422)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-forge/issues/327?issue=tenstorrent%7Ctt-forge%7C459)

### Problem description
In TT-XLA on the getting_started page, it's explained how to download a
wheel, but not how to get one that's pre-release.

### What's changed
Describe the approach used to solve the problem.
Provide a note that explains how users can use the --pre flag to grab a
pre-release wheel with the understanding that it may not be stable yet.

### Checklist
- [ ] New/Existing tests provide coverage for changes