Remove return_dict parametar from all model intializations (#159)
### Problem description

There was a mix for `return_dict` parameters, some models were setting
parameter to `False` which returns tuple. This makes testing training
workloads more difficult as we need to know for every model what is
order of returned tensors.

### What's changed
Removed `return_dict=` from all `from_pretrained` calls, as default
behavior of `return_dict=True` is preferred.

In codebase `return_dict` is only left in:
- `mplug_owl2/src/modeling_mplug_owl2.py` 
- `mplug_owl2/src/visual_encoder.py`
as this are actual implementations of model so it is not affecting.