Mark mutated buffers as input, non mutated buffers as constant (#1469)
### Ticket
N/A

### Problem description
- KV caches are marked as buffers by PyTorch. 
- We currently mark PyTorch buffers as "input" and do not consteval
them. This is because constevaling a KV cache tensor results in the
re-use of a cached KV cache tensor and produces poor data.
- Batchnorm mean/var are also marked as buffers, but we should consteval
these as they are not mutated and do not change between **inference**
iterations.

### What's changed
- If a buffer is mutated, then that input is also marked as a _bufffer
mutation_ in the ExportedProgram's `output_signature`.
- Using this information, we will now conditionally mark the buffer as
"input" if it is mutated, or "constant" if it is not. Thus allowing KV
cache tensors to persist while also constevaling batch norm mean/var.

### Checklist
- [X] New/Existing tests provide coverage for changes