Remove batchnorm decomposition (#1470)
* Decomposing batchnorm results in a graph with a lot more ops which
reduces performance.
* A better version of batchnorm decomposition (which allows for fusing
of the decomposed ops) exists in TTIR , so there is not need for it in
torch passes

Note: I'm not 100% sure how this will affect batchnorm for training, but
to my knowledge we are not training any models trough torch with
batchnorm currently. If this becomes a problem, support for lowering
training batchnorm to TTIR should be added.