Add BGE-M3 encode test (#1460)
### Ticket
#1196 

### Problem description
We want bge_m3 encode running in CI and recorded by superset. The
current TorchModelTester doesn't support inputs/outputs that aren't
tensors and so this model can't use the existing test infra.

### What's changed
Added a unique test for BAAI/bge-m3 encode_single_device. This test
loads the model from tt-forge-models, runs on cpu and device then
computes the PCC. PCC is currently ~~0.99~~ [EDIT]: _0.98 on bh_ and
asserted on so we can catch regressions. This test will run in the
nightly CI and report to superset.

Why TorchModelTester is not used: 
- The inputs are sentences which cannot be moved onto device using
`.to(device)`. In this test the inputs are moved onto device after
tokenization by passing in the device as an argument to
`encode_single_device`. This is why we use `encode_single_device`
instead of `encode` (encode doesn't take in a device argument)
- Some outputs aren't torch tensors and calls like `torch.allclose()`
fail to run when comparing them.
- `model.compile` doesn't support compiling `methods` but
`torch.compile` does

### Checklist
- [x] New/Existing tests provide coverage for changes