ModuleBuilder refactor to minimize state (#1328)
### Ticket
This PR is one of the prerequisites for #946 

This PR depends on my serialization branch and will be targetting it
until it is merged in main, at which point this PR will be rebased and
retargeted. Strictly speaking this PR does not depend on changes to
serialization, but a followup PR to this one will.

### Problem description
I didn't like that `ModuleBuilder` kept around state.

### What's changed
`ModuleBuilder` was refactored to not store results of `buildModule` as
state.
`ExecutableImage` was repurposed as a builder class for use inside
`ModuleBuilder`.
There was minor cleanup done to data being store, namely
`is_output_scalar` was removed.
Information about number of arguments is now extracted from IR and not
flatbuffer. Checks were added to ensure flatbuffer values match.

### Checklist
- [X] New/Existing tests provide coverage for changes