Add bge-m3 custom encode test (#1512)
### Ticket
#1196 

### Problem description
The encode function provided by BAAI/bge-m3 contains significant
overhead that is unnecessary and slows down the model. The demo in
tt-forge is being updated to use a custom encode function and this
custom function should be tested in tt-xla as well.

### What's changed
Added test_bge_m3_custom_encode with the custom encode function.

In addition to the expected encode outputs this function also returns
the outputs pre-processing so the PCC can be calculated from the device
outputs directly. Since this test compiles `bge_m3.model` and checks the
pcc of the `bge_m3.model` call it's recorded in superset as the
`bge_m3.model` test (tt-forge-models has `bge_m3.encode` which is
already tested in CI and `bge_m3.model` which is not yet tested in CI).

### Checklist
- [x] New/Existing tests provide coverage for changes