Add tt-mlir's venv to xla-ird docker image (#1527)
### Ticket
#1408 

### Problem description
The venv was removed from `/opt/ttmlir-toolchain` in the effort of
reducing Docker image size.
That caused the tt-xla-ird Docker image to be unable to build tt-mlir.

### What's changed
To reduce number of reservations on IRD and improve productivity, the
venv has been added back to tt-mlir toolchain dir.

Both CI and IRD variant of the Docker image are now built in
Dockerfile.ci.
The only difference is that the CI variant doesn't have the
`/opt/ttmlir-toolchain/venv` directory.

### Checklist
- [ ] New/Existing tests provide coverage for changes