Add decomposition for clamp const inputs (#1520)
### Problem description
Because torch constants created with `torch.ops.aten.scalar_tensor(x)`
(where x!=0 and x!=1) are lowered to vhlo as graph inputs we are not
able to perform fusing in mlir. This causes performance problems. The
issue for this is here:
https://github.com/tenstorrent/tt-xla/issues/1519

### What's changed
Add a decomposition in torch passes to represent integer constants as
additions for multiple ones. This gets optimized insto a single
`vhlo.constant` later.

### Checklist
- [ ] New/Existing tests provide coverage for changes

---------

Co-authored-by: Predrag Ilkic <148892209+pilkicTT@users.noreply.github.com>