Pre-commit, deprecation for tt-torch, fix doc link. (#1540)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-forge/issues/327?issue=tenstorrent%7Ctt-forge%7C486)

### Problem description
Provide context for the problem.
There is a broken link in the docs header, it goes into TT-XLA's folder
rather than the Tenstorrent doc site. TT-Torch is also not listed as
deprecated in the readme.

### What's changed
Describe the approach used to solve the problem.
Add the missing information and correct the docs link.

### Checklist
- [ ] New/Existing tests provide coverage for changes