[pjrt] input tensor persistence (#1523)
This change implements initial version of input tensor persistence, i.e.
leaving prepared tensors on the device for subsequent executions of the
same program. The goal being to minimize host->device transfers between
executions which is crucial for getting performance.

The `BufferInstance` now holds the host & prepared runtime tensors. The
`m_host_runtime_tensor` is the old `m_runtime_tensor` (just renamed it).

Whenever we prepare input tensors for program execution, they get saved
in the `BufferInstance` so that they can be reused on the subsequent
execution. The condition for reuse is that the already prepared tensor
is in the expected layout (from the flatbuffer binary) - this is needed
for cases when the same buffer instance is used in multiple programs
with different layouts.

The only complication is for sharded/replicated tensors; due to the
difference between PJRT and our runtime - we have only one tensor
object, while PJRT can have multiple buffer instances belonging to one
tensor. For now, we simply hold the prepared runtime tensor in all
buffer instances that represent it.

In future changes we can remove the host runtime tensors; I haven't done
it now to simplify this change.

The deallocation logic is modified to just reset the tensor object, and
let ref-counting do its magic.

NOTE: additionally the training test infra is fixed to run compiled
graphs.

Issue #953