Maximize space on runner when building tt-xla (#1565)
### Ticket
/

### Problem description
When building tt-xla in CI the runner (ubuntu-latest) is running out of
space.

### What's changed
A step that removed unnecessary files has been added to the `call-build`
workflow to increase the disk size before building tt-xla
This will allow the build for tt-xla to work in the workflows.

### Checklist
- [ ] New/Existing tests provide coverage for changes

### Validation
Workflow showing the change:
[here](https://github.com/tenstorrent/tt-xla/actions/runs/18219665072/job/51882254689)