Add torch_xla basic tests to onPR/onPush, return initialize_device_connectors, require tests to manually init xla computation client (#1550)
### Ticket
#1556

### What's changed
- return `initialize_device_connectors` autouse fixture
- remove `torch_xla._XLAC._init_computation_client()` call from
TorchDeviceConnector ctor, to avoid possible hangs when running pytorch
and jax tests in the same process.
- This means that tests will need to initialize the torch xla runtime
manually, and most already do so by acquiring a device. Maybe this can
be considered a reopening of #1383, since we don't have a good solution
to that yet
- Add push marks to torchxla basic and torchxla custom ops tests
- Due to --forked invocation ( #1554 ) add additional parallel splits to
these test jobs so they don't take 1.5 hours to run
- xfail 26 torchxla basic ops tests failing due to PCC errors with
unknown passing point #1555
- In TorchDeviceRunner only move workload.compiled_executable .to device
if .to is defined on the compiled_executable (not defined for customOps
OpTester tests)

### Checklist
- [ ] New/Existing tests provide coverage for changes