Reland "Split up ExecutableImage into FB and SO versions (#1329)" (#1606)
Original PR mangled some of the code in execution and as a result it was
as if the tensor persistence changes were reverted. We did not want that
so the PR was reverted. This PR relands the changes with the problem
fixed. Original description copy pasted below for posterity:

Ticket
This PR is one of the prerequisites for
https://github.com/tenstorrent/tt-xla/issues/946

This PR depends on https://github.com/tenstorrent/tt-xla/pull/1328 and
will be targetting it until it is merged in main, at which point this PR
will be rebased and retargeted.

Problem description
In preparation for Codegen(internally also called EmitC/EmitPy
sometimes) support, including using codegen for model execution, we
should split up classes to allow these separate code paths to exist.

What's changed
ExecutableImage and LoadedExecutableInstance have been split up.

Checklist
 New/Existing tests provide coverage for changes