Set mesh attr to shlo graph in spmd mode
We couldn't support fully replicated graph in torch-xla spmd mode, result to runtime error in torch_xla's pjrt client run as it degenerate mesh shape to [1, 1] which is unexpected.
This patch sets proper mesh shape attr to shlo graph to resolve this.
If sharding is marked by user or automatically, this will no-op so leave mesh shape as is.