Set mesh attr to shlo graph in spmd mode (#1560)
We couldn't support fully replicated graph in spmd mode, result to
runtime error in torch_xla's pjrt client run as it degenerate mesh shape
to [1, 1] which is unexpected. This patch sets proper mesh shape attr to
shlo graph to resolve this. If sharding is marked by user or
automatically, this will no-op so leave mesh shape as is.

### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-xla/issues/1487#issuecomment-3349604229)

### Problem description
In SPMD mode, we don’t currently support fully replicated graphs. This
results in a runtime error in torch_xla’s PJRT client because the mesh
shape degenerates to [1,1], which is unexpected.

### What's changed
- Normalize mesh shape in SPMD mode (uses `[1, num_devices]` instead of
`[1,1]` for fully replicated graphs).

### Checklist
- [v ] New/Existing tests provide coverage for changes