Add isort pre-commit plugin (#1614)
isort enforces python import sorting so this doesn't need to be checked
manually on PRs.

We have hundreds of files that don't abide by this convention and it
seems that adding a pre-commit hook triggers all files to be swept, so
fixes need to be made to all files by pre-commit run --all-files.