Codegen support, v1 (#1330)
### Ticket
Adresses core of #946 . However there will be more work left to be done

This PR depends on https://github.com/tenstorrent/tt-xla/pull/1329 and
will be targetting it until it is merged in main, at which point this PR
will be rebased and retargeted.

### Problem description
We want to support exporting models as C++ or Python TTNN code.

### What's changed
Support for codegen was added via tt-alchemist. This means models can be
exported as either C++ or Python using compile options. Actual execution
of models via C++ will come at a later date.

### Checklist
- [X] New/Existing tests provide coverage for changes