Enable JAX single-chip training tests (#1601)
### Ticket
Closes https://github.com/tenstorrent/tt-xla/issues/1388
Closes https://github.com/tenstorrent/tt-xla/issues/1610

### Problem description
TT-XLA currently only supports handful of training tests. The issues is
that most of the training tests are skipped as `Training support is not
implemented`. Apart from that there are some missing arguments in some
models (like MNIST with/without dropout) that need to be added in order
to run tests properly and get info on problems in TT repos.

### What's changed
This PR enables all single-chip training tests for JAX framework. 

Key additions of this PR are:

- Abstraction of `wrapper_model` in `jax_model_tester.py` to support
different type of outputs.
- Addition of pytest report tags for JAX single-chip training tests.
- Improvement of args/kwargs for specific training tests that do require
`train`, `rngs` etc. parameters to run. (Important change involves
setting train parameter by default. This change is going to fail models
that do not accept train parameter, which will make sure training test
addition is done in the proper way.)
- Replacement of `@pytest.mark.nightly` and `@pytest.mark.model_test`
for `@pytest.mark.training` for tests that do not pass, as we do not
want to run then in nightly CI.

This PR found out several new issues reported in following:
- https://github.com/tenstorrent/tt-mlir/issues/5304
- https://github.com/tenstorrent/tt-mlir/issues/5305
- https://github.com/tenstorrent/tt-mlir/issues/5307
- https://github.com/tenstorrent/tt-xla/issues/1650

This PR skips enablement of audio based models, due to problems with
setup:
- wav2vec2
- whisper
### Checklist
- [x] New/Existing tests provide coverage for changes