Add data parallel support for torch testing and initial set of 75 passing models (#1690)
### Ticket
N/A

### Problem description
We want to add support to runner/test_models.py and the
torch_model_tester infra so that single chip models can be tested in
data parallel (with batched inputs). This was supported in tt-torch.

### What's changed
- @kmabeeTT added `Parallelism.DATA_PARALLEL` as an option in
runner/test_config.py and added ~75 passing tests to CI
- added changes in runner/test_utils.py to batch inputs, create a shard
spec function and create a mesh when `self.parallelism ==
Parallelism.DATA_PARALLEL`
- added support in torch_device_runner.py for shard spec functions that
take in `args` and `kwargs`(data parallel) instead of `model` (tensor
parallel)

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Co-authored-by: Kyle Mabee <kmabee@tenstorrent.com>