[pjrt] add tracy zones
Tracy zones are added when compiled with `TTXLA_TRACY_ZONES=ON`.

All PJRT APIs are covered for now (except for error instance related
stuff), we can iterate on this if we notice that we are getting too
much noise; at this moment that doesn't seem to be the case.

Additionally, some expensive threads/methods are instrumented as well.

Cmake changes are to enable including needed `tracy` files. We will see
if it makes sense to install those in `tt-mlir` installation, so that we
can pick them up from there, but for now we take them from the
`tt-metal` path.

The way we set up include dirs is refactored, so that each lib defines
its own include dirs and propagates them via `PUBLIC` if needed.

Closes #3295