Tensor dumping (#1612)
### Ticket
One of the extended deliverables for #946 , also helps #1040. Closes
#1609 closes #933 .

### Problem description
We want a feature to enable our plugin to save all inputs to a graph
upon execution, in a format that is convenient for us to work with in
our tooling.
Update, this PR was extended to also include dumping IRs. Older approach
of relying on serialization had drawbacks.

### What's changed
A compile option `dump_inputs` was added that saves all inputs to a
graph to disk upon execution.
A compile optuon `dump_mlir_modules` was added that saves all IRs
produced by pipelines we invoke during compilation.

### Checklist
- [ ] New/Existing tests provide coverage for changes